4.21 done

+++ bloglist/controllers/blogs.js
@@ -55,9 +55,24 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  const blog = await Blog.findByIdAndDelete(request.params.id)
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+  const userid = user._id
+
+  const blog = await Blog.findById(request.params.id)
+  //const blog = await Blog.findByIdAndDelete(request.params.id)
   if (blog) {
-    response.status(204).end()
+    if ( blog.user.toString() === userid.toString() ) {
+      await Blog.findByIdAndDelete(request.params.id)
+      response.status(204).end()
+    }
+    else {
+      return response.status(401).json({ error: 'blog can only be deleted by the user who added it' })
+    }
   } else {
     response.status(404).end()
   }
